Add files via upload
The following "B" versions support ONLY Method 2 with the embedded Python interpreter. The mycMotDetRecPyB_config.ini is therefore stripped of some unused parameters! 
The difference here is that the Timing for recording is based on the parsed Time from the ONVIF XML messages send by the camera.
The examples above use for recording timing the clock of the PC/ Raspberry Pi 4.
mycMotDetRecPyB                  is the C++ program to start on Raspberry Pi 4 (64Bit Bookworm) in a terminal with ./mycMotDetRecPy
mycMotDetRecPyB_config.ini       contains all settings to control the program and access the camera stream and optional an AI Object Detection server
mycMotDetRecPyB.cpp              contains the C++ source code